pyproject.toml
[project]
name = "open-search-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "cyclopts>=5.0.0a1",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.2",
    "gitpython>=3.1.46",
    "opensearch-py>=3.1.0",
    "pygments>=2.19.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "tree-sitter==0.20.4",
    "tree-sitter-languages==1.10.2",
]
